rename: FilterConstants -> GatewayConstants 으로 수정
