+++ api-gateway/src/main/java/com/ohdelivery/gateway/apigateway/constants/GatewayConstants.java
@@ -4,7 +4,7 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class FilterConstants {
+public class GatewayConstants {
     public static final String AUTHORIZATION_HEADER = "Authorization";
     public static final String BEARER_PREFIX = "Bearer ";
     public static final String PASSPORT_ATTRIBUTE = "passport";
